did code cleanup, finally chose a data structure for the preprocessor, and also added a source file with input/output handling functions and implemented it  on the old code, and ran tests to make sure that it's working properly - everything seems great imo

+++ src/data_structures/binaryTree.c
@@ -74,6 +74,9 @@ void *treeSearch(TreeNode *root, char *key)
 
 static void *deleteTree_rec(TreeNode *root)
 {
+	if (!root)
+		return NULL;
+
 	if (root->leftChild != NULL)
 		root->leftChild = deleteTree_rec(root->leftChild);
